Support shareable type explorers
Some types have duplicate explorers. When one is updated, its
clones need to be tracked and updated as well. This commit is
a proof of concept allowing types to share explorers.

DESIGN RATIONALE

Explorers, whether __global or __type, are really just a tool to query
the state of an entity (be it a file, user list, package, mounts, etc.).
Having them stored in a common place makes sense not only from a
maintenance perspective, but also from a logical perspective.

The issue that emerges from grouping explorers in a common place is
having to figure out which must run before the initial manifest.
Putting them all together would require a registration system to
differentiate when they are designed to be used.

An alternative that we believe works better is having a simple
"type explorer library" located under __global/explorer. Types
would declare which shared explorers they depend on in a file so
they can be uploaded and run before running the type manifest.

This solution also provides backward-compatibilty, as cdist could
still handle explorers under __type/explorer.

PROOF OF CONCEPT

This commit introduces the notion of shared type explorers, located
under __global/explorer/type. A type may declare its dependency on
a shared explorer by listing it in __type/explorers.

Note that, in this PoC, if a shared explorer is listed, but missing,
cdist ignores the entry silently. It also ignores double entries,
where there exist a, say, "__type/explorer/myexplorer" and
"__global/explorer/type/myexplorer" (in the current implementation,
the former takes precedence).

FUTURE DIRECTIONS

- The dependency list could be named something else than "explorers"
  to prevent confusion.

- Performance has not been considered.